+++ src/main/java/com/tim/sernachat/SernaChatApplication.java
@@ -1,0 +1,13 @@
+package com.tim.sernachat;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SernaChatApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SernaChatApplication.class, args);
+    }
+
+}
