+++ src/main/java/com/tim/sernachat/controller/ChatController.java
@@ -1,0 +1,34 @@
+package com.tim.sernachat.controller;
+
+import com.tim.sernachat.model.ChatMessage;
+import com.tim.sernachat.repository.ChatRoomRepository;
+import com.tim.sernachat.service.ChatService;
+import com.tim.sernachat.service.JwtTokenProvider;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.messaging.handler.annotation.Header;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.stereotype.Controller;
+
+@Slf4j
+@RequiredArgsConstructor
+@Controller
+public class ChatController {
+
+    private final RedisTemplate<String, Object> redisTemplate;
+    private final JwtTokenProvider jwtTokenProvider;
+    private final ChatRoomRepository chatRoomRepository;
+    private final ChatService chatService;
+
+    /**
+     * websocket "/app/chat/message".
+     */
+    @MessageMapping("/chat/message")
+    public void message(ChatMessage message, @Header("token") String token) {
+        String nickname = jwtTokenProvider.getUserFromToken(token);
+        message.setFrom(nickname);
+        message.setUserCount(chatRoomRepository.getUserCount(message.getRoomId()));
+        chatService.sendChatMessage(message);
+    }
+}
