+++ src/main/java/com/tim/sernachat/controller/IndexController.java
@@ -1,0 +1,12 @@
+package com.tim.sernachat.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class IndexController {
+    @GetMapping({"", "/index"})
+    public String index() {
+        return "redirect:/chat/room";
+    }
+}
