+++ src/main/java/com/tim/sernachat/model/AuthRequest.java
@@ -1,0 +1,13 @@
+package com.tim.sernachat.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthRequest {
+    private String username;
+    private String password;
+}
