+++ src/main/java/com/tim/sernachat/model/ChatNotification.java
@@ -1,0 +1,7 @@
+package com.tim.sernachat.model;
+
+public class ChatNotification {
+    private String id;
+    private String fromId;
+    private String fromName;
+}
