+++ src/main/java/com/tim/sernachat/model/LoginInfo.java
@@ -1,0 +1,16 @@
+package com.tim.sernachat.model;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class LoginInfo {
+    private final String name;
+    private final String token;
+
+    @Builder
+    public LoginInfo(String name, String token) {
+        this.name = name;
+        this.token = token;
+    }
+}
