+++ src/main/java/com/tim/sernachat/model/MessageStatus.java
@@ -1,0 +1,5 @@
+package com.tim.sernachat.model;
+
+public enum MessageStatus {
+    RECEIVED, DELIVERED
+}
