+++ src/main/java/com/tim/sernachat/model/Role.java
@@ -1,0 +1,5 @@
+package com.tim.sernachat.model;
+
+public enum Role {
+    ROLE_USER, ROLE_ADMIN, ROLE_GUEST
+}
