+++ src/main/java/com/tim/sernachat/repository/ChatHistoryRepository.java
@@ -1,0 +1,7 @@
+package com.tim.sernachat.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ChatHistoryRepository {
+}
