+++ src/main/java/com/tim/sernachat/repository/ChatRoomRepository.java
@@ -1,0 +1,64 @@
+package com.tim.sernachat.repository;
+
+import com.tim.sernachat.model.ChatRoom;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.HashOperations;
+import org.springframework.data.redis.core.ValueOperations;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+import javax.annotation.Resource;
+
+@RequiredArgsConstructor
+@Repository
+public class ChatRoomRepository {
+    public static final String USER_COUNT = "USER_COUNT";
+    public static final String ENTER_INFO = "ENTER_INFO";
+    // Redis CacheKeys
+    private static final String CHAT_ROOMS = "CHAT_ROOM";
+    @Resource(name = "redisTemplate")
+    private HashOperations<String, String, ChatRoom> hashOpsChatRoom;
+    @Resource(name = "redisTemplate")
+    private HashOperations<String, String, String> hashOpsEnterInfo;
+    @Resource(name = "redisTemplate")
+    private ValueOperations<String, String> valueOps;
+
+    public ChatRoom createChatRoom(String name) {
+        ChatRoom chatRoom = ChatRoom.create(name);
+        hashOpsChatRoom.put(CHAT_ROOMS, chatRoom.getRoomId(), chatRoom);
+        return chatRoom;
+    }
+
+    public List<ChatRoom> findAllRoom() {
+        return hashOpsChatRoom.values(CHAT_ROOMS);
+    }
+
+    public ChatRoom findRoomById(String id) {
+        return hashOpsChatRoom.get(CHAT_ROOMS, id);
+    }
+
+    public void setUserEnterInfo(String sessionId, String roomId) {
+        hashOpsEnterInfo.put(ENTER_INFO, sessionId, roomId);
+    }
+
+    public String getUserEnterRoomId(String sessionId) {
+        return hashOpsEnterInfo.get(ENTER_INFO, sessionId);
+    }
+
+    public void removeUserEnterInfo(String sessionId) {
+        hashOpsEnterInfo.delete(ENTER_INFO, sessionId);
+    }
+
+    public long getUserCount(String roomId) {
+        return Long.valueOf(Optional.ofNullable(valueOps.get(USER_COUNT + "_" + roomId)).orElse("0"));
+    }
+
+    public long plusUserCount(String roomId) {
+        return Optional.ofNullable(valueOps.increment(USER_COUNT + "_" + roomId)).orElse(0L);
+    }
+
+    public long minusUserCount(String roomId) {
+        return Optional.ofNullable(valueOps.decrement(USER_COUNT + "_" + roomId)).filter(count -> count > 0).orElse(0L);
+    }
+}
