+++ src/main/java/com/tim/sernachat/service/ChatService.java
@@ -1,0 +1,37 @@
+package com.tim.sernachat.service;
+
+import com.tim.sernachat.model.ChatMessage;
+import com.tim.sernachat.repository.ChatRoomRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ChatService {
+    private final ChannelTopic channelTopic;
+    private final RedisTemplate redisTemplate;
+    private final ChatRoomRepository chatRoomRepository;
+
+
+    public String getRoomId(String destination) {
+        int lastIndex = destination.lastIndexOf('/');
+        if (lastIndex != -1)
+            return destination.substring(lastIndex + 1);
+        else
+            return "";
+    }
+
+    public void sendChatMessage(ChatMessage chatMessage) {
+        chatMessage.setUserCount(chatRoomRepository.getUserCount(chatMessage.getRoomId()));
+        if (ChatMessage.MessageType.ENTER.equals(chatMessage.getType())) {
+            chatMessage.setMessage(chatMessage.getFrom() + "— Вы вошли в комнату");
+            chatMessage.setFrom("[уведомление]");
+        } else if (ChatMessage.MessageType.QUIT.equals(chatMessage.getType())) {
+            chatMessage.setMessage(chatMessage.getFrom() + "— Вы вошли в комнату");
+            chatMessage.setFrom("[уведомление]");
+        }
+        redisTemplate.convertAndSend(channelTopic.getTopic(), chatMessage);
+    }
+}
