--- src/main/java/com/timurisachenko/websocketchat/SpringWebsocketChatApplication.java
@@ -1,13 +1,0 @@
-package com.timurisachenko.websocketchat;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringWebsocketChatApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringWebsocketChatApplication.class, args);
-    }
-
-}
