--- src/main/java/com/timurisachenko/websocketchat/config/EmbeddedRedisConfig.java
@@ -1,35 +1,0 @@
-package com.timurisachenko.websocketchat.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import redis.embedded.RedisServer;
-
-import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
-
-@Profile("local")
-@Configuration
-public class EmbeddedRedisConfig {
-
-    @Value("${spring.redis.host:127.0.0.1}")
-    private String redisHost;
-
-    @Value("${spring.redis.port:6379}")
-    private int redisPort;
-
-    private RedisServer redisServer;
-
-    @PostConstruct
-    public void redisServer() {
-        redisServer = new RedisServer(redisPort);
-        redisServer.start();
-    }
-
-    @PreDestroy
-    public void stopRedis() {
-        if (redisServer != null) {
-            redisServer.stop();
-        }
-    }
-}
