--- src/main/java/com/timurisachenko/websocketchat/controller/IndexController.java
@@ -1,12 +1,0 @@
-package com.timurisachenko.websocketchat.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class IndexController {
-    @GetMapping({"", "/index"})
-    public String index() {
-        return "redirect:/chat/room";
-    }
-}
