--- src/main/java/com/timurisachenko/websocketchat/model/AuthResponse.java
@@ -1,13 +1,0 @@
-package com.timurisachenko.websocketchat.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthResponse {
-    private String token;
-
-}
