--- src/main/java/com/timurisachenko/websocketchat/model/ChatNotification.java
@@ -1,7 +1,0 @@
-package com.timurisachenko.websocketchat.model;
-
-public class ChatNotification {
-    private String id;
-    private String fromId;
-    private String fromName;
-}
