--- src/main/java/com/timurisachenko/websocketchat/model/ChatRoom.java
@@ -1,26 +1,0 @@
-package com.timurisachenko.websocketchat.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-@Getter
-@Setter
-public class ChatRoom implements Serializable {
-    private static final long serialVersionUID = 6494678977089006639L;
-
-    private String roomId;
-    private String name;
-    private long userCount;
-
-    public static ChatRoom create(String name) {
-        ChatRoom chatRoom = new ChatRoom();
-        chatRoom.roomId = UUID.randomUUID().toString();
-        chatRoom.name = name;
-        return chatRoom;
-    }
-}
-
-
