--- src/main/java/com/timurisachenko/websocketchat/model/LoginInfo.java
@@ -1,16 +1,0 @@
-package com.timurisachenko.websocketchat.model;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class LoginInfo {
-    private final String name;
-    private final String token;
-
-    @Builder
-    public LoginInfo(String name, String token) {
-        this.name = name;
-        this.token = token;
-    }
-}
