--- src/main/java/com/timurisachenko/websocketchat/model/MessageStatus.java
@@ -1,5 +1,0 @@
-package com.timurisachenko.websocketchat.model;
-
-public enum MessageStatus {
-    RECEIVED, DELIVERED
-}
