--- src/main/java/com/timurisachenko/websocketchat/model/Role.java
@@ -1,5 +1,0 @@
-package com.timurisachenko.websocketchat.model;
-
-public enum Role {
-    ROLE_USER, ROLE_ADMIN, ROLE_GUEST
-}
