--- src/main/java/com/timurisachenko/websocketchat/repository/ChatHistoryRepository.java
@@ -1,7 +1,0 @@
-package com.timurisachenko.websocketchat.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChatHistoryRepository {
-}
