+++ src/test/java/com/tim/sernachat/SernaChatApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tim.sernachat;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SernaChatApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
