--- src/test/java/com/timurisachenko/websocketchat/SpringWebsocketChatApplicationTests.java
@@ -1,13 +1,0 @@
-package com.timurisachenko.websocketchat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringWebsocketChatApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
